Fixed JS bug preventing removal of Gource options; support additive options
- Fixed DOM selection bug that listed Build options + Project options
  when saving Gource settings (preventing ability to remove entries)
- Added `ProjectOption.unique_together` for `(project, name)`
- Added `ProjectBuildOption.unique_together` for `(build, name)`
- Changed API to add/update `gource_options` dictionary vs. sync
  - Added `sync_gource_options=True` flag in request to perform sync of settings
- Updated API unit tests for Gource options

+++ gource_studio/gource_studio/core/api/views.py
@@ -295,10 +295,21 @@ class ProjectDetail(ProjectPermissionQuerySetMixin, generics.RetrieveUpdateDestr
                         response['gource_options'] = f"Invalid Gource option '{option}'"
                         return Response(response, status=status.HTTP_400_BAD_REQUEST)
 
-                # Delete old options
-                project.options.all().delete()
-                # Add new set
-                ProjectOption.objects.bulk_create(new_options)
+                # Check for flag to indicate if old options should be removed
+                sync_options = request.data.get('sync_gource_options', False)
+                if sync_options:
+                    # Delete old options
+                    project.options.all().delete()
+                    # Add new set
+                    ProjectOption.objects.bulk_create(new_options)
+                else:
+                    # Merge with current set
+                    ProjectOption.objects.bulk_create(
+                        new_options,
+                        update_conflicts=True,
+                        unique_fields=['project', 'name'],
+                        update_fields=['project', 'name']
+                    )
             else:
                 logging.error(f"Invalid 'gource_options' provided: {request.data['gource_options']}")
                 response['gource_options'] = "Invalid Gource options type (must be a dict)"
